Move client/consensus state types
