+++ crates/chain/chain-type-components/src/traits/ibc/client_state.rs
@@ -1,0 +1,9 @@
+use cgp::prelude::*;
+
+#[derive_component(ClientStateTypeComponent, ProvideClientStateType<Chain>)]
+pub trait HasClientStateType<Counterparty>: Async {
+    /**
+        The client state of the `Self` chain's client on the `Counterparty` chain
+    */
+    type ClientState: Async;
+}
