+++ crates/chain/chain-type-components/src/traits/ibc/consensus_state.rs
@@ -1,0 +1,9 @@
+use cgp::prelude::*;
+
+#[derive_component(ConsensusStateTypeComponent, ProvideConsensusStateType<Chain>)]
+pub trait HasConsensusStateType<Counterparty>: Async {
+    /**
+        The consensus state of the `Self` chain's client on the `Counterparty` chain
+    */
+    type ConsensusState: Async;
+}
