+++ crates/chain/chain-type-components/src/traits/ibc/mod.rs
@@ -1,6 +1,8 @@
 pub mod channel_id;
 pub mod client_id;
+pub mod client_state;
 pub mod connection_id;
+pub mod consensus_state;
 pub mod packet;
 pub mod port_id;
 pub mod sequence;
